+++ src/main/java/org/n52/supervisor/api/CheckTask.java
@@ -13,18 +13,18 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+package org.n52.supervisor.api;
 
-package org.n52.supervisor.tasks;
+import java.util.Collection;
 
-import org.n52.supervisor.CheckRunner;
 
 /**
  * 
  * @author Daniel
  * 
  */
-public interface CheckTaskFactory {
+public interface CheckTask {
 
-    CheckTask create(CheckRunner checker);
+    public Collection<CheckResult> checkIt(CheckRunner c);
 
 }
