+++ src/main/java/org/n52/supervisor/api/CheckTaskFactory.java
@@ -14,28 +14,16 @@
  * limitations under the License.
  */
 
-package org.n52.supervisor.checks.util;
+package org.n52.supervisor.api;
 
-import javax.xml.bind.annotation.XmlRootElement;
-
-import org.n52.supervisor.checks.Check;
 
 /**
  * 
  * @author Daniel
  * 
  */
-@XmlRootElement
-public class CheckCollection extends Check {
+public interface CheckTaskFactory {
 
-    protected String type = "CheckCollection";
-
-    public CheckCollection() {
-        super();
-    }
-
-    public CheckCollection(String identifier) {
-        super(identifier);
-    }
+    CheckTask create(CheckRunner checker);
 
 }
