+++ src/main/java/org/n52/supervisor/api/Notification.java
@@ -14,29 +14,17 @@
  * limitations under the License.
  */
 
-package org.n52.supervisor;
+package org.n52.supervisor.api;
 
 import java.util.Collection;
 
-import org.n52.supervisor.checks.Check;
 
 /**
- * 
  * @author Daniel Nüst (d.nuest@52north.org)
  * 
  */
-public interface CheckFactory {
+public interface Notification {
 
-    public static final long EVERY_12_HOURS = 1000 * 60 * 60 * 12;
-
-    public static final long EVERY_24_HOURS = 1000 * 60 * 60 * 24;
-
-    public static final long EVERY_HALF_HOUR = 1000 * 60 * 30;
-
-    public static final long EVERY_HOUR = 1000 * 60 * 60;
-
-    public static final long EVERY_WEEK = 1000 * 60 * 60 * 24 * 7;
-
-    public abstract Collection<Check> getChecks();
+    public Collection<CheckResult> getResults();
 
 }
