+++ src/main/java/org/n52/supervisor/checks/ows/SosLatestObservationCheckRunner.java
@@ -37,8 +37,8 @@ import net.opengis.swe.x101.TimeObjectPropertyType;
 
 import org.apache.xmlbeans.XmlCursor;
 import org.apache.xmlbeans.XmlObject;
+import org.n52.supervisor.api.CheckResult;
 import org.n52.supervisor.checks.AbstractServiceCheckRunner;
-import org.n52.supervisor.checks.CheckResult;
 import org.n52.supervisor.checks.ServiceCheckResult;
 import org.n52.supervisor.util.XmlTools;
 import org.slf4j.Logger;
