+++ src/main/java/org/n52/supervisor/checks/util/SelfCheckRunner.java
@@ -20,14 +20,14 @@ import java.util.Collection;
 import java.util.Date;
 
 import org.n52.supervisor.SupervisorInit;
+import org.n52.supervisor.api.Check;
+import org.n52.supervisor.api.CheckResult;
+import org.n52.supervisor.api.Notification;
+import org.n52.supervisor.api.UnsupportedCheckException;
+import org.n52.supervisor.api.CheckResult.ResultType;
 import org.n52.supervisor.checks.AbstractServiceCheckRunner;
-import org.n52.supervisor.checks.Check;
-import org.n52.supervisor.checks.CheckResult;
-import org.n52.supervisor.checks.CheckResult.ResultType;
-import org.n52.supervisor.checks.UnsupportedCheckException;
 import org.n52.supervisor.db.ResultDatabase;
-import org.n52.supervisor.ui.EmailNotification;
-import org.n52.supervisor.ui.Notification;
+import org.n52.supervisor.notification.EmailNotification;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
