+++ src/main/java/org/n52/supervisor/db/ResultDatabase.java
@@ -23,7 +23,7 @@ import java.util.List;
 import java.util.Queue;
 import java.util.concurrent.LinkedBlockingQueue;
 
-import org.n52.supervisor.checks.CheckResult;
+import org.n52.supervisor.api.CheckResult;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
