+++ src/main/java/org/n52/supervisor/resources/Checks.java
@@ -40,9 +40,9 @@ import javax.ws.rs.core.UriInfo;
 import org.codehaus.jettison.json.JSONArray;
 import org.codehaus.jettison.json.JSONException;
 import org.codehaus.jettison.json.JSONObject;
-import org.n52.supervisor.CheckRunner;
 import org.n52.supervisor.CheckerResolver;
-import org.n52.supervisor.checks.Check;
+import org.n52.supervisor.api.Check;
+import org.n52.supervisor.api.CheckRunner;
 import org.n52.supervisor.db.CheckDatabase;
 import org.n52.supervisor.db.ResultDatabase;
 import org.n52.supervisor.tasks.ManualChecker;
